Add files via upload
The adc_testrgb.py file reads the analog readings from the frequency-to-analog voltage output and changes the color based on the changes in the analog input.

The RGB_all_combinations.py file detects if the four pads are touched and returns them as boolean constants. The main part checks all combinations of the four touch pads and produces an oscillating output.